requirements.txt
python-dotenv==1.0.0
spotipy==2.23.0
yt-dlp==2023.9.24